requirements.txt
appdirs==1.4.3
packaging==16.8
pyparsing==2.2.0
pyvmomi==6.5
PyYAML==3.12
requests==2.14.2
six==1.10.0
xmltodict==0.11.0
